[Add] ObjectSystem

+++ Server/ECS/Component/Components.cs
@@ -6,6 +6,6 @@ public record struct VelocityComponent(float vx, float vy);
 
 public record struct BoundColliderComponent(float cx, float cy, float w, float h);
 
-public record struct PacketSendTimer(float sendTime, float elasedTime);
+public record struct PacketSendTimer(double sendTime, double elasedTime);
 
 public record struct EntityTypeComponent(EEntityType type);
